Remove JSDOM tests and replace with Puppeteer tests
JSDOM was great to get us started, but it doesn't work for plugins
that have to deal with the real page (navigate, observe AJAX
requests and DOM changes, etc.). This impacts especially Quora,
which is our most used plugin and which wasn't testable to date!

Puppeteer is a nice way to automate Chromium within Node. We've
put this in a new `cli` directory in anticipation of future tools
like a command-line grabber.

Still very experimental but all tests appear to work now!

A couple of changes along the way:
- made DataSet more Node-friendly
- gave retrieval functions an URL option, in case we want to
  call them with an explicit profile URL (which the CLI supports)
- updated docs
- updated jQuery